+++ src/anomalib/models/__init__.py
@@ -15,6 +15,7 @@ from anomalib.utils.path import convert_to_snake_case
 from .image import (
     Cfa,
     Cflow,
+    ChatGPTVision,
     Csflow,
     Dfkde,
     Dfm,
@@ -24,7 +25,6 @@ from .image import (
     Fastflow,
     Fre,
     Ganomaly,
-    OpenaiVlm,
     Padim,
     Patchcore,
     ReverseDistillation,
@@ -43,6 +43,7 @@ class UnknownModelError(ModuleNotFoundError):
 __all__ = [
     "Cfa",
     "Cflow",
+    "ChatGPTVision",
     "Csflow",
     "Dfkde",
     "Dfm",
@@ -52,7 +53,6 @@ __all__ = [
     "Fastflow",
     "Fre",
     "Ganomaly",
-    "OpenaiVlm",
     "Padim",
     "Patchcore",
     "ReverseDistillation",
