+++ src/anomalib/models/image/chatgpt_vision/__init__.py
@@ -3,6 +3,6 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .lightning_model import OpenaiVlm
+from .lightning_model import ChatGPTVision
 
-__all__ = ["OpenaiVlm"]
+__all__ = ["ChatGPTVision"]
